rename surface.py -> main.py and move TotalErosion -> erosion.py
