+++ fastscape/processes/main.py
@@ -24,12 +24,12 @@ class TotalVerticalMotion:
     bedrock = xs.variable(
         dims=('y', 'x'),
         intent='out',
-        description='total bedrock motion in upward direction'
+        description='bedrock motion in upward direction'
     )
     surface = xs.variable(
         dims=('y', 'x'),
         intent='out',
-        description='total topographic surface motion in downward direction'
+        description='topographic surface motion in upward direction'
     )
 
     def run_step(self):
@@ -141,41 +141,6 @@ class UniformSoilLayer:
 
     def run_step(self):
         self.thickness = self._get_thickness()
-
-
-@xs.process
-class TotalErosion:
-    """Sum up all erosion processes."""
-
-    erosion_vars = xs.group('erosion')
-
-    cumulative_erosion = xs.variable(
-        dims=[(), ('y', 'x')],
-        intent='inout'
-    )
-
-    erosion = xs.variable(
-        dims=[(), ('y', 'x')],
-        intent='out',
-        description='total erosion',
-        group='surface_downward'
-    )
-
-    erosion_rate = xs.on_demand(
-        dims=[(), ('y', 'x')],
-        description='erosion rate (all processes)'
-    )
-
-    @xs.runtime(args='step_delta')
-    def run_step(self, dt):
-        self._dt = dt
-
-        self.erosion = np.sum((err for err in self.erosion_vars))
-        self.cumulative_erosion += self.erosion
-
-    @erosion_rate.compute
-    def _erosion_rate(self):
-        return self.erosion / self._dt
 
 
 @xs.process
